requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
pytesseract==0.3.13
Pillow==11.0.0
google-generativeai==0.8.3
dateparser==1.2.0
python-dotenv==1.0.1
httpx==0.27.2
